[util] Move http-related function to http.go
Along with related tests. In order to avoid putting everything
in the util/common.go file.
