+++ pkg/util/common.go
@@ -6,22 +6,14 @@
 package util
 
 import (
-	"crypto/tls"
-	"fmt"
 	"io"
 	"io/ioutil"
 	"math/rand"
-	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/util/log"
-
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
-	"github.com/DataDog/datadog-agent/pkg/config"
-	"github.com/DataDog/datadog-agent/pkg/version"
 )
 
 func init() {
@@ -97,16 +89,6 @@ func EnsureParentDirsExist(p string) error {
 	return nil
 }
 
-// HTTPHeaders returns a http headers including various basic information (User-Agent, Content-Type...).
-func HTTPHeaders() map[string]string {
-	av, _ := version.New(version.AgentVersion, version.Commit)
-	return map[string]string{
-		"User-Agent":   fmt.Sprintf("Datadog Agent/%s", av.GetNumber()),
-		"Content-Type": "application/x-www-form-urlencoded",
-		"Accept":       "text/html, */*",
-	}
-}
-
 // GetJSONSerializableMap returns a JSON serializable map from a raw map
 func GetJSONSerializableMap(m interface{}) interface{} {
 	switch x := m.(type) {
@@ -139,71 +121,4 @@ func GetJSONSerializableMap(m interface{}) interface{} {
 	}
 	return m
 
-}
-
-// GetProxyTransportFunc return a proxy function for a http.Transport that
-// would return the right proxy depending on the configuration.
-func GetProxyTransportFunc(p *config.Proxy) func(*http.Request) (*url.URL, error) {
-	return func(r *http.Request) (*url.URL, error) {
-		// check no_proxy list first
-		for _, host := range p.NoProxy {
-			if r.URL.Host == host {
-				log.Debugf("URL match no_proxy list item '%s': not using any proxy", host)
-				return nil, nil
-			}
-		}
-
-		// check proxy by scheme
-		confProxy := ""
-		if r.URL.Scheme == "http" {
-			confProxy = p.HTTP
-		} else if r.URL.Scheme == "https" {
-			confProxy = p.HTTPS
-		} else {
-			log.Warnf("Proxy configuration do not support scheme '%s'", r.URL.Scheme)
-		}
-
-		if confProxy != "" {
-			proxyURL, err := url.Parse(confProxy)
-			if err != nil {
-				err := fmt.Errorf("Could not parse the proxy URL for scheme %s from configuration: %s", r.URL.Scheme, err)
-				log.Error(err.Error())
-				return nil, err
-			}
-			userInfo := ""
-			if proxyURL.User != nil {
-				if _, isSet := proxyURL.User.Password(); isSet {
-					userInfo = "*****:*****@"
-				} else {
-					userInfo = "*****@"
-				}
-			}
-
-			log.Debugf("Using proxy %s://%s%s for URL '%s'", proxyURL.Scheme, userInfo, proxyURL.Host, SanitizeURL(r.URL.String()))
-			return proxyURL, nil
-		}
-
-		// no proxy set for this request
-		return nil, nil
-	}
-}
-
-// CreateHTTPTransport creates an *http.Transport for use in the agent
-func CreateHTTPTransport() *http.Transport {
-	tlsConfig := &tls.Config{
-		InsecureSkipVerify: config.Datadog.GetBool("skip_ssl_validation"),
-	}
-
-	if config.Datadog.GetBool("force_tls_12") {
-		tlsConfig.MinVersion = tls.VersionTLS12
-	}
-
-	transport := &http.Transport{
-		TLSClientConfig: tlsConfig,
-	}
-
-	if proxies := config.GetProxies(); proxies != nil {
-		transport.Proxy = GetProxyTransportFunc(proxies)
-	}
-	return transport
 }
